Mail from Edward:
I've taken a look at this demo, and the problem here is that DE
doesn't actually implement a fixed point semantics.  This demo would
be a good test case for a re-implementation of DE Director derived from
FixedPointDirector.

In my view, it was an accident that this demo ever worked.
Can we promote the demo to a "known failed test"?

Edward

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@54244 3435fffc-ccdd-42ab-8df9-e21c6b74a643