Fixed AtomicActor_Limiter test case to work with TesterDirector
The original test seemed to rely on pruning the "ineffective"
constraints away, in that it claimed that the unconnected ports
should resolve to null rather than "Unknown."  I feel more confident
with Unknown (which is the current behavior with the TesterDirector),
so I've changed the golden output.

Also, the original test harness, and double clicking on the Solver
did not (and still don't) work in this case.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@56377 3435fffc-ccdd-42ab-8df9-e21c6b74a643