Added first take on the InterfaceCheckerDirector.
This director tries to make sure that all the contained actors
express their interfaces through special parameters.

If the actor has a parameter named _interfaceExpr, this is
interpreted as a boolean-valued Ptolemy expression representing
the interface.

Otherwise, if there is a parameter named _interfaceStr, this
is interpreted as a string in the Yices expression language
representing the interface.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@56489 3435fffc-ccdd-42ab-8df9-e21c6b74a643