Fixed definitions of test/train to work with null concepts.
This is needed as the solver currently can output null for certain
conceptables that are not considered part of the current inference.
A possible alternate approach would be to have a function that only
returns a list of the "active" conceptables, rather than looping
through all of them, like we do here.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@57526 3435fffc-ccdd-42ab-8df9-e21c6b74a643