Added getExecutionIdentifier() and setExecutionIdentifier(). See http://bugzilla.ecoinformatics.org/show_bug.cgi?id=4230
"In SDF provenance records to the error table an exception
thrown after the workflow stops. In PN, multiple exceptions
may be thrown, and this happens during execution. Right now
there's a problem with these errors getting
recorded. Related/in addition: there's also difficulty
knowing which execution errors are correlated with, since
the same workflow may be executing and throwing errors in
multiple windows. As Dan notes: Manager reports exceptions
asynchronously; need a way to tie an exception to
corresponding execution."

"The issue is that the Manager notifies listeners of model
errors asynchronously to avoid deadlocks.  However, this
makes it impossible for a listener to associate a specific
execution cycle with a given error."


git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@57596 3435fffc-ccdd-42ab-8df9-e21c6b74a643