Added functionality for comparing Concepts with inequalities.
This makes the test added in the previous commit pass, but it's a little
ugly, since it involves very similar code as in the base class
implementataion.  A better approach may be to make a special interface
for partial orders that provides functionality for equality and
inequality comparison, and then just make sure that both ScalarTokens
and Concepts fulfill that interface.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@57914 3435fffc-ccdd-42ab-8df9-e21c6b74a643