Added test that has different domain.
In this test, the expression is over the dimension
lattice rather that over the montonicity lattice.
I think this is a more realistic use of the
monotonicity lattice.  Unfortunately, it doesn't
yet work, since the inference assumes that all
expressions are over the lattice that the analysis
is over.  Correcting this assumption will be the
subject of the following commit.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@58618 3435fffc-ccdd-42ab-8df9-e21c6b74a643