Added false positive case.
This is a monotonic function, but the simplest if
analysis that doesn't consider the constant in the
conditional can't realize that and will consider
it nonmonotonic.  This is the simplest possible
such case, since there is no border at all between
the if and else branches (the if is statically
true).  Even so, I intend to use an analysis
similar to the ifc analysis of the paper, rather
than some type of constant propagation.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@58621 3435fffc-ccdd-42ab-8df9-e21c6b74a643