Factored IfNodeFunction into its own class
It was by far the most complicated part of the
ASTPTFunctionalIfNode adapter, so I thought it
deserved to be more than an inner class.  I'm
not sure that the current location is correct,
though, since I think that this folder is
supposed to only contain things that could be
adapters, which ConceptFunctions are not.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@58810 3435fffc-ccdd-42ab-8df9-e21c6b74a643