Extended analysis to work on integrator case.
This extension extends the analysis to remember
counterexamples at one level of the AST and
remember that the next level of the AST is in
fact monotonic if it is monotonic at those
counterexamples.

This overall approach is exactly what I want,
but the particular implementation is not ideal.
It saves the counterexamples in a global hash
table, similar to how the concepts are stored
during resolution.  A better approach would
be to have infinite concepts who can contain
the counterexamples as data, or somehow store
the counterexamples as attributes on Ptolemy
AST nodes

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@58815 3435fffc-ccdd-42ab-8df9-e21c6b74a643