Hi Chris, Your changes look good.
One comment is that your method comments should not be longer than 80
characters per line so that if the code is printed, it looks ok.  I realize
that this is in contradiction to the edict that we not use Eclipse to
indent comments automatically.

Why don't you run the comment indenter on the basic/layout/kieler package
to clean the method comments up and then I'll fix up the other
////////////////////////////////////////////////////////
////             public methods                     ////

comments.

_Christopher 

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59279 3435fffc-ccdd-42ab-8df9-e21c6b74a643