ConceptTokens can now wrap any Concept.
This includes infinite concepts. This requires
a cascade of changes to other classes that
unbox Concept values from ConceptTokens.  This
fixes all the  issues that prevent compilation,
but leaves a few warts that require fixing:

1. The implementation irresponsibly allows
cases where the Ontology's leastUpperBound
method is called, even though it will not
work in cases that include InfiniteConcepts.

2. More problematically, the monotonicity
analysis example is completely broken for now,
unable to infer over conditional statements.
A fix will be included in a follwing patch,
but this will require a more substantial
overhau.  The new monotonicity analysis will
need to use InfiniteConcepts to allow
asserting the monotonicity of expressions
with multiple free variables, such as
  (x <= Const) ? Unused :
  (y <= Unused) ? Nonconst :
  Const
(over the Const-Nonconst lattice).
This example should have a monotonicity of
{x:Monotonic, y:Antimonotonic} in the new
infinite montonicity lattice.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59491 3435fffc-ccdd-42ab-8df9-e21c6b74a643