Simplified the multiplication rule.
I think that the bottom element should take
precedence over zero, in that multiplying
zero by the empty set of elements should
give no result rather than a zero result.
For example, if we take the result of multiplying
by zero and multiply that by another zero,
I would expect that we should still get an error,
rather than a zero-valued result.

For now, this breaks the sclae demo, which relys
on this behavior, but I'll either change the demo
or the constraints in a following commit.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59495 3435fffc-ccdd-42ab-8df9-e21c6b74a643