Abstracted automatic concept name generation.
Now this feature resides in InfiniteConcept rather
than in MonotonicityConcept.  This allows infinite
concepts to be created with just an associated
ontology and no name.  This is useful in cases
like record concepts where the toString and equals
methods are determined by the content of the
concept and not of the name.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59567 3435fffc-ccdd-42ab-8df9-e21c6b74a643