Added test for new monotoncity analysis.
This is pretty much the simplest possible test
case, testing an expressiong with a single leaf
node with a constant value. In the new, infinite
monotonicity lattice, the inferred concept should
not be "Constant," however.  Rather, it should be
a hash of all variables for which the monotonicity
is non-constant.  In the case of a completely
constant expression, this will just be the empty
hash.

This case does not yet pass, but will be made to
pass in the subsequent commits.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59666 3435fffc-ccdd-42ab-8df9-e21c6b74a643