Make equals handle ontologyless InfiniteConcepts.
I'm not sure how an InfiniteConcept can ever be
created without an ontology (this _should_ be
impossible), but a concept without an ontology
should not be equal to any other concept.

I should also go back soon and verify how it
is possible that InfiniteConcepts are being
created without ontologies.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59668 3435fffc-ccdd-42ab-8df9-e21c6b74a643