Made interfaces of concept graphs more concrete.
Since Concept <= Object, this still fulfills the
contract of CPO, but knowing that these methods
return a Concept makes their method signatures
more informative.  The next step is to create a
common supertype/interface of ConceptGraph and
ProductLatticeCPO so that other classes can rely
on the fact that these things return Concepts.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59694 3435fffc-ccdd-42ab-8df9-e21c6b74a643