Let leaf nodes also support domain lattice.
I don't think that this approach is the right way
to support them, as an exception is still thrown
during type checking when run on a single leaf
node with a lattice constant in it.  Still, this
may be enough to get all of the conditional node
tests to pass.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59753 3435fffc-ccdd-42ab-8df9-e21c6b74a643