Fixed analysis to work over non-"x" varibales.
This simply loops over all the variables used in
the expression doing each one individually, so
cases which involve evaluation (like the if
cases) will not yet work.  For those to work, we
would need to modify the analysis to take into
account having multiple free variables within a
signle branch of an if statement. This is future
work.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59782 3435fffc-ccdd-42ab-8df9-e21c6b74a643