Stop inferring concepts for expression inputs.
I thought that the fact that they were coceptable
made the resulting output more confusing overall,
so now I've removed them for clarity.

A future extension would be to see if the
Expression adapter could be removed completely,
and have the functionality be included in an
ActorConstraintsDefinition that has the input set
to IGNORE and the output >= expression.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59811 3435fffc-ccdd-42ab-8df9-e21c6b74a643