Change "strict dominators" to "cover set."
I think this terminology is more standard in order
theory, and it's closer to the terminology used by
Davey and Priestley. I also needed a name for the
reverse also (the elements directly below a given
concept), so I just went with "cover set below."
Uninventive, but at least clear.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59825 3435fffc-ccdd-42ab-8df9-e21c6b74a643