More info to retrun type of getLexicalTokenList()
This makes some things cleaner, and should work in
all existing cases.  I debated having it return a
List<? extends Token> for compatibility, but it
seems that nothing relies on this list being of a
type that can contain only certain Token subtypes,
so I'll leave it as simple as possible.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@59999 3435fffc-ccdd-42ab-8df9-e21c6b74a643