Test case using exact constraints from MODELS.
Some of the syntax needed to be munged a bit to
make the monotonicity analysis work on the
constraints, but they are semantically exactly the
MODELS constaints, and syntactically about as
close as we can get with the current monotonicity
analyzer.  If we add projections to the expression
evaluator and monotonicity analysis (projections
are always monotonic), then we could make the
syntax a bit cleaner.  Even in that case, it
wouldn't be perfectly the same syntax as in the
MODELS paper, as that uses x == y in a
conditional, which wouldn't really be possible to
analyze without a quite major extension to the
monotonicity analysis.

git-svn-id: svn+ssh://source.eecs.berkeley.edu/chess/ptII/trunk@60259 3435fffc-ccdd-42ab-8df9-e21c6b74a643